Ahn's WinApi Lecture 0428

+++ WinApi/Win32_Test_Framework/TimeSystem.h
@@ -8,6 +8,6 @@ namespace time
 
 	const float GetFrameRate();
 
-	const ULONGLONG GetDeltaTime();
+	const float GetDeltaTime();
 
 }
